requirements.txt
flask
-e git+https://github.com/mark-burnett/jot.git#egg=jot
oauthlib
pyyaml
requests
requests-oauthlib
simplejson
sqlalchemy
